checking merge file name is already present or not.

+++ phy-merge-pdf-tool/pythonscript/mergefiles.py
@@ -1,4 +1,6 @@
+
 import PyPDF2 
+import os
 
 def merge1(pdfiles, mergepdf_filename):
     pdfMerge = PyPDF2.PdfMerger()
@@ -9,10 +11,18 @@ def merge1(pdfiles, mergepdf_filename):
         pdf_reader = PyPDF2.PdfReader(pdfFile)
         pdfMerge.append(pdf_reader)
         pdfFile.close()
+    
 
-    with open(directory + mergepdf_filename +'.pdf', 'wb') as mergedFile:
-        pdfMerge.write(mergedFile)
-#C://Users//User//Documents// +{mergepdf_filename} +.pdf
+    if (directory + mergepdf_filename +'.pdf') == os.path.isfile(directory + mergepdf_filename +'.pdf'):
+        print("true")
+
+        with open(directory + mergepdf_filename +'.pdf', 'wb') as mergedFile:
+            pdfMerge.write(mergedFile)
+    
+    else: 
+        print("false")
+
+    #C://Users//User//Documents// +{mergepdf_filename} +.pdf
     print("Successfully merged pdf files.")
     #return "Merged file is created in " + directory + mergepdf_filename +'.pdf';
